fix: VIA Timer 1/2 cycle counter overflow at 1x speed
T1LastTime and T2LastTime were uint32_t but the cycle counter
(ScaledCycleCount) is uint64_t.  After ~8.6s at 1x speed the counter
crossed 2^32, causing deltaTime to wrap to ~4 billion, which made the
timer think it had expired.  The interrupt handler would fire, return
via RTE, and immediately re-enter — an infinite loop that froze the
emulator.

Also swap T2CL read order: call doTimer2Check() before
clrInterruptFlag() to prevent the check from immediately re-raising
the flag that was just cleared.

[skip ci]

+++ src/devices/via_base.h
@@ -73,10 +73,10 @@ public:
 	uint8_t T2_Active = 0;
 	bool T1IntReady = false;
 	bool T1Running = true;
-	uint32_t T1LastTime = 0;
+	uint64_t T1LastTime = 0;
 	bool T2Running = true;
 	bool T2C_ShortTime = false;
-	uint32_t T2LastTime = 0;
+	uint64_t T2LastTime = 0;
 
 protected:
 	int viaNum_;			// 1 or 2
